techfuturist-app/requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
email-validator
